Updating script/push_contributing_docuemnt.rb
This script will now create a branch and make a pull request for
updating or creating the CONTRIBUTING.md document.

Note: It has been tested in creating a branch, but has not been tested
for creating the related pull request (I didn't think about doing that
before I ran update with a new branch).

But don't worry, we can test this thing against a change related to our
squash policy.